requirements.txt
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
lightgbm==4.3.0
paho-mqtt==1.6.1
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0
joblib==1.3.2
